+++ examples/vue-kitchen-sink/src/stories/core.stories.js
@@ -13,7 +13,7 @@ export default {
   },
 };
 
-export const PassedToStory = ({ parameters: { fileName, ...parameters } }) => ({
+export const PassedToStory = (_args, { parameters: { fileName, ...parameters } }) => ({
   template: `<div>Parameters are <pre>${JSON.stringify(parameters, null, 2)}</pre></div>`,
 });