+++ lib/client-api/src/client_api.test.ts
@@ -116,7 +116,7 @@ describe('preview.client_api', () => {
       const { storiesOf } = clientApi;
 
       clientApi.addParameters({ a: 1 });
-      storiesOf('kind', module).add('name', ({ parameters }) => parameters);
+      storiesOf('kind', module).add('name', (_args, { parameters }) => parameters);
 
       const result = storyStore.fromId('kind--name').storyFn();
       // @ts-ignore
@@ -131,7 +131,7 @@ describe('preview.client_api', () => {
 
       clientApi.addParameters({ options: { a: '1' } });
       clientApi.addParameters({ options: { b: '2' } });
-      storiesOf('kind', module).add('name', ({ parameters }) => parameters);
+      storiesOf('kind', module).add('name', (_args, { parameters }) => parameters);
 
       // @ts-ignore
       const {
@@ -147,7 +147,7 @@ describe('preview.client_api', () => {
 
       clientApi.addParameters({ backgrounds: ['value'], options: { a: '1', b: '3' } });
       clientApi.addParameters({ options: { a: '2' } });
-      storiesOf('kind', module).add('name', ({ parameters }) => parameters);
+      storiesOf('kind', module).add('name', (_args, { parameters }) => parameters);
 
       // @ts-ignore
       const {
@@ -165,7 +165,7 @@ describe('preview.client_api', () => {
 
       clientApi.addParameters({ backgrounds: ['value'], options: { a: '1', b: '3' } });
       clientApi.addParameters({ backgrounds: [], options: { a: '2' } });
-      storiesOf('kind', module).add('name', ({ parameters }) => parameters);
+      storiesOf('kind', module).add('name', (_args, { parameters }) => parameters);
 
       // @ts-ignore
       const {
@@ -183,7 +183,7 @@ describe('preview.client_api', () => {
 
       clientApi.addParameters({ options: { a: '1', b: '2', theming: { c: '3' } } });
       clientApi.addParameters({ options: { theming: { c: '4', d: '5' } } });
-      storiesOf('kind', module).add('name', ({ parameters }) => parameters);
+      storiesOf('kind', module).add('name', (_args, { parameters }) => parameters);
 
       // @ts-ignore
       const {
@@ -245,7 +245,7 @@ describe('preview.client_api', () => {
 
       storiesOf('kind', module)
         .addDecorator((fn) => `aa-${fn()}`)
-        .add('name', (c) => `${c.kind}-${c.name}`);
+        .add('name', (_args, c) => `${c.kind}-${c.name}`);
 
       const result = storyStore.fromId('kind--name').storyFn();
       expect(result).toBe(`aa-kind-name`);